cafegory/requirements.txt
django
Pillow
jinja2 
termcolor
uwsgi
django-allauth
django-star-ratings
